Todos/Findings: ##### todos long term: make a boundary for pressure variance and start testing with an object in a fluid investigate the separation of data in multiple points -> separation of wet and dry nodes -> separation of data-array form the nodes -> new class for the macro values rework toLinks, is mostly in order needs more testing thou make wierd channels to should be m4 for me -> next step calculate seems superior than the lookup thou, prob comes down to cpu stalls, should also do a cache hit analysis prob to shed some light on that -> still test out with more tables prob needs a SoA structure to work porperlly thou so more a bigger thing than a smaller... # idas deleting the neighborhood list after generation seems somewhat unproductive alternatively keep it and instead of doing a xy search do a search in the form of the velocity set force based pressure boundary?! ##### fixed the interaction with the new calcualte straightintersection funcion

+++ src/straight.h
@@ -52,7 +52,7 @@ class straightGenerator {
     void surface_mass_center();
     double calculate_total_surface_length();
     double calculate_total_surface_length(boundaryType_t type);
-    bool calculate_straight_intersection(straight_t* to_be_check,straight_t * reference,point_t* intersection_point, int* used_vector_length);
+    bool calculate_straight_intersection(straight_t* to_be_check,straight_t * reference,point_t* intersection_point, double* used_vector_length);
 };
 
 bool straight_better_candidate_test(straight_t* candidate, straight* partner);
